refactored map model and controller
